Refactor all the way config file is loaded (now by each plugin). Should solve issues #1508.
