+++ conf/glances.conf
@@ -188,15 +188,17 @@ critical=90
 #folder_3_path=/nonexisting
 #folder_4_path=/root
 
+[hddtemp]
+# Define hddtemp server IP and port (default is 127.0.0.1 and 7634 (TCP))
+host=127.0.0.1
+port=7634
+
 [sensors]
 # Sensors core thresholds (in Celsius...)
 # Default values if not defined: 60/70/80
 temperature_core_careful=60
 temperature_core_warning=70
 temperature_core_critical=80
-# Define hddtemp server IP and port (default is 127.0.0.1 and 7634 (TCP))
-hddtemp_host=127.0.0.1
-hddtemp_port=7634
 # Temperatures threshold in °C for hddtemp
 # Default values if not defined: 45/52/60
 temperature_hdd_careful=45
