+++ glances/plugins/glances_alert.py
@@ -88,9 +88,10 @@ class Plugin(GlancesPlugin):
     Only for display.
     """
 
-    def __init__(self, args=None):
+    def __init__(self, args=None, config=None):
         """Init the plugin."""
         super(Plugin, self).__init__(args=args,
+                                     config=config,
                                      stats_init_value=[])
 
         # We want to display the stat in the curse interface
