+++ glances/plugins/glances_docker.py
@@ -72,9 +72,10 @@ class Plugin(GlancesPlugin):
     stats is a dict: {'version': {...}, 'containers': [{}, {}]}
     """
 
-    def __init__(self, args=None):
+    def __init__(self, args=None, config=None):
         """Init the plugin."""
         super(Plugin, self).__init__(args=args,
+                                     config=config,
                                      items_history_list=items_history_list)
 
         # The plgin can be disable using: args.disable_docker
