+++ glances/plugins/glances_irq.py
@@ -33,9 +33,10 @@ class Plugin(GlancesPlugin):
     stats is a list
     """
 
-    def __init__(self, args=None):
+    def __init__(self, args=None, config=None):
         """Init the plugin."""
         super(Plugin, self).__init__(args=args,
+                                     config=config,
                                      stats_init_value=[])
 
         # We want to display the stat in the curse interface
