+++ glances/plugins/glances_processlist.py
@@ -91,9 +91,10 @@ class Plugin(GlancesPlugin):
         'name': '[{}]'
     }
 
-    def __init__(self, args=None):
+    def __init__(self, args=None, config=None):
         """Init the plugin."""
         super(Plugin, self).__init__(args=args,
+                                     config=config,
                                      stats_init_value=[])
 
         # We want to display the stat in the curse interface
