+++ glances/plugins/glances_quicklook.py
@@ -59,9 +59,10 @@ class Plugin(GlancesPlugin):
     'stats' is a dictionary.
     """
 
-    def __init__(self, args=None):
+    def __init__(self, args=None, config=None):
         """Init the quicklook plugin."""
         super(Plugin, self).__init__(args=args,
+                                     config=config,
                                      items_history_list=items_history_list)
         # We want to display the stat in the curse interface
         self.display_curse = True
