+++ glances/plugins/glances_raid.py
@@ -39,9 +39,9 @@ class Plugin(GlancesPlugin):
     stats is a dict (see pymdstat documentation)
     """
 
-    def __init__(self, args=None):
+    def __init__(self, args=None, config=None):
         """Init the plugin."""
-        super(Plugin, self).__init__(args=args)
+        super(Plugin, self).__init__(args=args, config=config)
 
         # We want to display the stat in the curse interface
         self.display_curse = True
