+++ glances/plugins/glances_wifi.py
@@ -49,9 +49,10 @@ class Plugin(GlancesPlugin):
     Get stats of the current Wifi hotspots.
     """
 
-    def __init__(self, args=None):
+    def __init__(self, args=None, config=None):
         """Init the plugin."""
         super(Plugin, self).__init__(args=args,
+                                     config=config,
                                      stats_init_value=[])
 
         # We want to display the stat in the curse interface
